refactor dateUtils
