+++ packages/frontend/src/mobile-components/notifications/Notifications.tsx
@@ -1,9 +1,9 @@
 import { FC, MouseEventHandler, useState } from "react";
 import { twMerge } from "@alphaday/ui-kit";
 import { EFeedItemType } from "src/api/types";
+import { computeDuration } from "src/api/utils/dateUtils";
 import { ReactComponent as BellSVG } from "src/assets/icons/bell.svg";
 import { ReactComponent as CloseSVG } from "src/assets/icons/close2.svg";
-import { computeDuration } from "src/utils/dateUtils";
 import { getFeedItemIcon } from "../feed/FeedElements";
 
 type TNotification = {
